[feat] 아카데미 출입증 인식 뷰 Service Error Alert 추가 및 서버 API 호출 실패 시 알러트 로직 추가
아카데미 출입증 인식 뷰 Service Error Alert 추가 및 서버 API 호출 실패 시 알러트 띄워주는 로직을 추가하였습니다.
